backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
dataclasses==0.6
Django==4.2.20
django-cloudinary-storage==0.3.0
django-mongoengine==0.5.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.7.0
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mongoengine==0.29.1
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.0b1
pydantic_core==2.31.1
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
